refactor: 기본 타입들 리팩터링
