+++ frontend/src/types/index.ts
@@ -4,13 +4,12 @@ export interface UserType {
   profileUrl: string;
 }
 
-export type ImageSizeType = 'EXTRA_HUGE' | 'HUGE' | 'LARGE' | 'MEDIUM' | 'SMALL' | 'TINY';
-
-export interface RequestType {
-  accessToken: string | null;
-  method: 'get' | 'post' | 'put' | 'delete';
-  url: string;
-  headers: {
-    Authorization: string;
-  };
+export interface ParticipantType {
+  memberId: string;
+  levellogId: string;
+  preQuestionId: string;
+  nickname: string;
+  profileUrl: string;
 }
+
+export type ImageSizeType = 'EXTRA_HUGE' | 'HUGE' | 'LARGE' | 'MEDIUM' | 'SMALL' | 'TINY';
