--- frontend/src/types/member.ts
@@ -1,5 +1,0 @@
-export interface MemberType {
-  id: string;
-  nickname: string;
-  profileUrl: string;
-}
